backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
